dom/credentialmanagement/identity/tests/mochitest/test_well_known_multiple_config_simple.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Happypath Test for multiple possible configs in the well known</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    SpecialPowers.pushPrefEnv({ set: [["dom.security.credentialmanagement.identity.test_well_known_path", "/tests/dom/credentialmanagement/identity/tests/mochitest/web-identity-multiple-config" ]] }).then(() => {
      setupTest("simple")
    }).then(
      function () {
        SpecialPowers.wrap(document).notifyUserGestureActivation();
        return navigator.credentials.get({
          identity: {
            mode: "active",
            providers: [{
              configURL: "https://example.net/tests/dom/credentialmanagement/identity/tests/mochitest/server_manifest.sjs",
              clientId: "mochitest",
              nonce: "nonce"
            }]
          }
        });
      }
    ).then((cred) => {
      ok(true, "successfully got a credential");
      is(cred.token,
        "account_id=1234&client_id=mochitest&nonce=nonce&disclosure_text_shown=false&is_auto_selected=false",
        "Correct token on the credential.");
      is(cred.id,
        "1234",
        "Correct id on the credential");
      is(cred.type,
        "identity",
        "Correct type on the credential");
    }).catch(() => {
      ok(false, "must not have an error");
    }).finally(async () => {
      await SpecialPowers.popPrefEnv();
      SimpleTest.finish();
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">This makes sure we can still get a credential when there is more than one entry in `provider_urls` in our well-known (`./web-identity-multiple-config`)</div>
<pre id="test"></pre>
</body>
</html>
